Move kart/schema.py to kart/tabular
