+++ kart/text_diff_writer.py
@@ -10,7 +10,7 @@ from .base_diff_writer import BaseDiffWriter
 
 from .output_util import format_wkt_for_output, resolve_output_path
 from .feature_output import feature_as_text, feature_field_as_text
-from .schema import Schema
+from .tabular.schema import Schema
 
 
 _NULL = object()
