+++ kart/working_copy/postgis.py
@@ -12,7 +12,7 @@ from sqlalchemy.orm import sessionmaker
 from .db_server import DatabaseServer_WorkingCopy
 from .table_defs import PostgisKartTables
 from kart import crs_util
-from kart.schema import Schema
+from kart.tabular.schema import Schema
 from kart.sqlalchemy import separate_last_path_part
 from kart.sqlalchemy.adapter.postgis import KartAdapter_Postgis
