+++ kart/working_copy/sqlserver.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import sessionmaker
 from .db_server import DatabaseServer_WorkingCopy
 from .table_defs import SqlServerKartTables
 from kart import crs_util
-from kart.schema import Schema
+from kart.tabular.schema import Schema
 from kart.sqlalchemy import separate_last_path_part, text_with_inlined_params
 from kart.sqlalchemy.adapter.sqlserver import KartAdapter_SqlServer
