+++ tests/test_working_copy.py
@@ -1,4 +1,4 @@
-from kart.schema import ALL_DATA_TYPES
+from kart.tabular.schema import ALL_DATA_TYPES
 
 _ALL_SUBTYPES = {
     "integer": (8, 16, 32, 64),
